Move cJSON to external
