+++ rtengine/procparams.cc
@@ -23,6 +23,7 @@
 
 #include <locale.h>
 
+#include <cjson/cJSON.h>
 #include <glib/gstdio.h>
 #include <glibmm/fileutils.h>
 #include <glibmm/keyfile.h>
@@ -31,7 +32,6 @@
 #include "params/serdes.h"
 
 #include "aspectratios.h"
-#include "cJSON.h"
 #include "color.h"
 #include "colortemp.h"
 #include "curves.h"
